Moved some utils directly to 110
